Player movement via nodes
Nodes have been placed throughout the maze so that character movement is by navigating a graph rather than through collision detection. The player is able to control PacMan to mooth smoothly between nodes and navigate the maze.

The project has implemented up to part 5 of this tutorial https://youtu.be/TyQLweQzhlM